+++ public_html/php/modify_businesses.php
@@ -8,7 +8,7 @@
  * Used in: js\settings.js
  *
  * Created on Apr 06, 2024
- * Updated on Apr 12, 2024
+ * Updated on Jul 25, 2024
  *
  * Description: Check if the user is signed in and modify the tbl_businesses table.
  * Dependenties: config.php
@@ -222,7 +222,7 @@ function CheckBusinesses($id, $group, $business)
  * Function:    DeleteBusinesses
  *
  * Created on Apr 07, 2024
- * Updated on Apr 07, 2024
+ * Updated on Jul 25, 2024
  *
  * Description: Delete the row with id in the tbl_businesses table.
  *
@@ -232,26 +232,73 @@ function CheckBusinesses($id, $group, $business)
  */    
  function DeleteBusinesses($id)
  {   
-    $response = [];  
-       
+    $response = CheckBusinessInFinances($id);
+    if ($response['success'] && !$response['exists'])
+    { 
+        try 
+        {    
+            $db = OpenDatabase();
+                
+            $query = "DELETE FROM tbl_businesses WHERE `id` = $id";          
+            $select = $db->prepare($query);
+            $select->execute();
+                    
+            $response['success'] = true;  
+        }
+        catch (PDOException $e) 
+        {    
+            $response['message'] = $e->getMessage();
+            $response['success'] = false;
+        } 
+    }
+    
+    // Close database connection.
+    $db = null;   
+      
+    return $response;
+}
+
+/*
+ * Function:    CheckBusinessInFinances
+ *
+ * Created on Jul 25, 2024
+ * Updated on Jul 25, 2024
+ *
+ * Description: Check if the business exists in the tbl_finances table.
+ *
+ * In:  $id
+ * Out: $response
+ *
+ */
+function CheckBusinessInFinances($id)
+{  
+    $response = [];
+    
     try 
-    {    
+    { 
         $db = OpenDatabase();
                 
-        $query = "DELETE FROM tbl_businesses WHERE `id` = $id";          
+        // Check if the business exists in the tbl_finances table.
+        $query = "SELECT count(0) FROM tbl_finances WHERE bid = $id;";        
         $select = $db->prepare($query);
-        $select->execute();
-                    
-        $response['success'] = true;  
-    }
+        $select->execute();        
+        $result = $select->fetchColumn();
+
+        $response['exists'] = false; 
+        if ($result > 0) {
+            $response['exists'] = true; 
+        }  
+
+        $response['success'] = true;         
+    }    
     catch (PDOException $e) 
     {    
         $response['message'] = $e->getMessage();
         $response['success'] = false;
     } 
-
+    
     // Close database connection.
-    $db = null;   
-      
+    $db = null;        
+    
     return $response;
 }
